Issue 场景ui开发

See merge request ccpg-hotel/construction-android!84

+++ app/src/main/java/com/ayla/hotelsaas/mvp/present/OneKeyPresenter.java
@@ -12,7 +12,7 @@ import io.reactivex.schedulers.Schedulers;
 
 public class OneKeyPresenter extends BasePresenter<OneKeyView> {
 
-    public void runRuleEngine(int ruleId) {
+    public void runRuleEngine(long ruleId) {
         RequestModel.getInstance().runRuleEngine(ruleId)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
